+++ src/derivatives_trading_portfolio_margin/rest_api/models/margin_max_borrow_response.rs
@@ -18,9 +18,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct MarginMaxBorrowResponse {
     #[serde(rename = "amount", skip_serializing_if = "Option::is_none")]
-    pub amount: Option<rust_decimal::Decimal>,
+    pub amount: Option<String>,
     #[serde(rename = "borrowLimit", skip_serializing_if = "Option::is_none")]
-    pub borrow_limit: Option<i64>,
+    pub borrow_limit: Option<String>,
 }
 
 impl MarginMaxBorrowResponse {